% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFeaturingNetwork.R
\name{CreateFeaturingNetwork}
\alias{CreateFeaturingNetwork}
\title{Featuring network from artists songs}
\usage{
CreateFeaturingNetwork(x)
}
\arguments{
\item{x}{A dataset coming from scraping function of genius API}
}
\value{
2 dataframes : One with a Source node and Target node, the other containing infos about artists. Those dataframe can be exported in csv and used in a network analysing software like Gephi
}
\description{
Create a network of featuring collaboration between artists
}
\examples{

\dontrun{

## Get a network of the following df :
feat_net<-CreateFeaturingNetwork(dataframe)
# Export the network to your computer and remove the first column (rownames)
write.csv(feat_net, "/PATH_TO_DIRECTORY/users_network.csv", rownames = FALSE)

}

}
