% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSongsArtistsID.R
\name{GetSongsArtistsID}
\alias{GetSongsArtistsID}
\title{Scrap Artists and songs information from ID}
\usage{
GetSongsArtistsID(x)
}
\arguments{
\item{x}{A list of artists IDs}
}
\value{
A dataframe
}
\description{
This function allow you to get the information aboute artists and songs from an artist ID from Genius API. You need to have package geniusr installed
}
\examples{

\dontrun{

## Get a dataset from the following list :
ID_list <- c("ID-1", "ID-2")
# Use the Function to retrieve information
artists_info <- GetSongsArtistsID(ID_list)

}

}
