% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSongsArtistsName.R
\name{GetSongsArtistsName}
\alias{GetSongsArtistsName}
\title{Scrap Artists and songs information}
\usage{
GetSongsArtistsName(x)
}
\arguments{
\item{x}{A list of artists names}
}
\value{
A dataframe
}
\description{
This function allow you to get the information about artists and songs from Genius API. You need to have package geniusr installed
}
\examples{

\dontrun{

## Get a dataset from the following list :
artists_list <- c("Bruce Springsteen", "Pink Floyd")
# Use the Function to retrieve information
artists_info <- GetSongsArtistsName(artists_list)

}

}
