% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveSongs.R
\name{RetrieveSongs}
\alias{RetrieveSongs}
\title{Featuring network from artists songs}
\usage{
RetrieveSongs(x, y, feat_exclude = T)
}
\arguments{
\item{x}{A dataset containing at least 3 columns : artist_id, song_name, and song_id (you can use the output of scrap_songs_artist or scrap_songs_artist_from_ID)}

\item{y}{A list of Artists IDS from which you want to scrap all the lyrics}

\item{feat_exclude}{if False, the function will get all the songs of the artist even the featuring with other artists. Default is set to True (for personnal purpose) but it could be pertinent to set it to false}
}
\value{
a dataframes with all the lyrics, with columns about artists, song etc...
}
\description{
Create a network of featuring collaboration between artists. The recuperation can be pretty long if there is a lot of song. For instance, to get all the lyrics from 10 artists it would take approximatly 1 hour
}
\examples{

\dontrun{

## all the lyrics of multiple artists :
lyrics<-RetreiveSongs(corpus,list_of_artists_id,feat_exclude=F)

}

}
