% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_songs_artists.R
\name{scrap_songs_artists}
\alias{scrap_songs_artists}
\title{Scrap Artists and songs information}
\usage{
scrap_songs_artists(x)
}
\arguments{
\item{x}{A list of artists names}
}
\value{
A dataframe
}
\description{
This function allow you to get the information aboute artists and songs from Genius API. You need to have package geniusr installed
}
\examples{

\dontrun{

## Get a dataset from the following list :
artists_list <- c("artist-1", "artist-2")
# Use the Function to retrieve information
artists_info <- scrap_songs_artists(artists_list)

}

}
