% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_featuring_dataset_list.R
\name{write_featuring_dataset_list}
\alias{write_featuring_dataset_list}
\title{Write dataset list}
\usage{
write_featuring_dataset_list(dataframes_list)
}
\arguments{
\item{dataframes_list}{The result of the featuring network function. A variable consisting in a list of two dataframes : one for the nodes of the network and the other for the edges}
}
\value{
The csv files are exported in the working directory and can be then used in a network analysing software like Gephi
}
\description{
Exports the results of the featuring network function function under 2 csv file in your working directory (!! May not yet work with windows)
}
\examples{

\dontrun{

## Get a network of featuring between artists :
network <- create_featuring_network(base)
# Export the network to your computer and remove the first column (rownames)
write_featuring_dataset_list<-(network)

}

}
